+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Add PersonRecord Component
 - Remove uri prop compulsion in Avatar Component
 - Add disabled prop in Link Component
+- Solved Swipe use in people list item
+- Fix people list swipe issue
 
 ---